+++ analysis/decompile/ida/functions/00445840-kill_subgoldy.c
@@ -5,6 +5,6 @@
 // Void `Player` member that hides both cached ghost sprites and clears Goldy's active render flag. Its sole `update_times_up` caller discards the final flag-clear residue.
 void __thiscall kill_subgoldy(Player *player)
 {
-  begin_post_follow_carryover((int)player);
+  begin_post_follow_carryover(player);
   player->body.transform.position.y = -8.0;
 }
